requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
pymysql==1.1.0
cryptography==41.0.3
